Refactor camera models into self-contained sensor/models/ headers
Split the 3145-line sensor/models.h into an umbrella header plus a new
sensor/models/ folder: base.h (id enum, macros, CRTP bases), one header
per model family, runtime.h (CameraModel* dispatch), and jacobian.h
(moved from models_jacobian.h, no content change).

Define all model implementations inline in the class bodies instead of
in out-of-line sections, and slim the definition macros accordingly.

Derive the parameter-group index arrays from the group sizes passed to
the macros via a constexpr helper, pass params_info through the macros,
and initialize model_name/params_info/idx members directly (constexpr
where possible). This removes all Initialize*() boilerplate methods and
their models.cc definitions.

Replace the name/id lookup maps with a jump-table switch for id->name
and a lazily-built O(1) map for name->id. Derive the fisheye predicate
from the base class instead of a hand-maintained model list. Rename
CamRayFromImgJacobian to CamRayFromImgJac.

No behavior change: all index arrays, params info strings, and name/id
round trips verified against the previous implementation.
