+++ src/colmap/sensor/models.cc
@@ -33,90 +33,41 @@
 
 namespace colmap {
 
-// Define the static members shared by all camera models.
-#define CAMERA_MODEL_CASE(CameraModel)           \
-  constexpr CameraModelId CameraModel::model_id; \
-  const std::string CameraModel::model_name =    \
-      CameraModel::InitializeModelName();        \
-  constexpr size_t CameraModel::num_params;      \
-  const std::string CameraModel::params_info =   \
-      CameraModel::InitializeParamsInfo();
-
-CAMERA_MODEL_CASES
-
-#undef CAMERA_MODEL_CASE
-
-// Define the parameter-group members specific to perspective models.
-#define CAMERA_MODEL_CASE(CameraModel)                    \
-  const std::array<size_t, CameraModel::num_focal_params> \
-      CameraModel::focal_length_idxs =                    \
-          CameraModel::InitializeFocalLengthIdxs();       \
-  const std::array<size_t, CameraModel::num_pp_params>    \
-      CameraModel::principal_point_idxs =                 \
-          CameraModel::InitializePrincipalPointIdxs();    \
-  const std::array<size_t, CameraModel::num_extra_params> \
-      CameraModel::extra_params_idxs =                    \
-          CameraModel::InitializeExtraParamsIdxs();
-
-PERSPECTIVE_CAMERA_MODEL_CASES
-
-#undef CAMERA_MODEL_CASE
-
-// Define the parameter-group members specific to spherical models.
-#define CAMERA_MODEL_CASE(CameraModel)                       \
-  const std::array<size_t, CameraModel::num_metadata_params> \
-      CameraModel::metadata_idxs =                           \
-          CameraModel::InitializeMetaDataParamsIdxs();
-
-SPHERICAL_CAMERA_MODEL_CASES
-
-#undef CAMERA_MODEL_CASE
-
-NodeHashMap<std::string, CameraModelId> InitialzeCameraModelNameToId() {
-  NodeHashMap<std::string, CameraModelId> camera_model_name_to_id;
-
-#define CAMERA_MODEL_CASE(CameraModel)                     \
-  camera_model_name_to_id.emplace(CameraModel::model_name, \
-                                  CameraModel::model_id);
-
-  CAMERA_MODEL_CASES
-
-#undef CAMERA_MODEL_CASE
-
-  return camera_model_name_to_id;
-}
-
-NodeHashMap<CameraModelId, const std::string*> InitialzeCameraModelIdToName() {
-  NodeHashMap<CameraModelId, const std::string*> camera_model_id_to_name;
-
-#define CAMERA_MODEL_CASE(CameraModel)                   \
-  camera_model_id_to_name.emplace(CameraModel::model_id, \
-                                  &CameraModel::model_name);
-
-  CAMERA_MODEL_CASES
-
-#undef CAMERA_MODEL_CASE
-
-  return camera_model_id_to_name;
-}
-
-static const NodeHashMap<std::string, CameraModelId> kCameraModelNameToId =
-    InitialzeCameraModelNameToId();
-
-static const NodeHashMap<CameraModelId, const std::string*>
-    kCameraModelIdToName = InitialzeCameraModelIdToName();
-
 bool ExistsCameraModelWithName(const std::string& model_name) {
-  return kCameraModelNameToId.count(model_name) > 0;
+  return CameraModelNameToId(model_name) != CameraModelId::kInvalid;
 }
 
 bool ExistsCameraModelWithId(const CameraModelId model_id) {
-  return kCameraModelIdToName.count(model_id) > 0;
+  switch (model_id) {
+#define CAMERA_MODEL_CASE(CameraModel) \
+  case CameraModel::model_id:          \
+    return true;
+
+    CAMERA_MODEL_CASES
+
+#undef CAMERA_MODEL_CASE
+    default:
+      return false;
+  }
 }
 
 CameraModelId CameraModelNameToId(const std::string& model_name) {
-  const auto it = kCameraModelNameToId.find(model_name);
-  if (it == kCameraModelNameToId.end()) {
+  // Function-local static: built once on first use (thread-safe), keeping
+  // the name lookup O(1) without paying for eager static initialization.
+  static const NodeHashMap<std::string, CameraModelId> kNameToId = [] {
+    NodeHashMap<std::string, CameraModelId> name_to_id;
+#define CAMERA_MODEL_CASE(CameraModel) \
+  name_to_id.emplace(CameraModel::model_name, CameraModel::model_id);
+
+    CAMERA_MODEL_CASES
+
+#undef CAMERA_MODEL_CASE
+
+    return name_to_id;
+  }();
+
+  const auto it = kNameToId.find(model_name);
+  if (it == kNameToId.end()) {
     return CameraModelId::kInvalid;
   } else {
     return it->second;
@@ -124,13 +75,20 @@ CameraModelId CameraModelNameToId(const std::string& model_name) {
 }
 
 const std::string& CameraModelIdToName(const CameraModelId model_id) {
-  const auto it = kCameraModelIdToName.find(model_id);
-  if (it == kCameraModelIdToName.end()) {
-    const static std::string kEmptyModelName = "";
-    return kEmptyModelName;
-  } else {
-    return *(it->second);
+  switch (model_id) {
+#define CAMERA_MODEL_CASE(CameraModel) \
+  case CameraModel::model_id:          \
+    return CameraModel::model_name;
+
+    CAMERA_MODEL_CASES
+
+#undef CAMERA_MODEL_CASE
+    default:
+      break;
   }
+
+  const static std::string kEmptyModelName = "";
+  return kEmptyModelName;
 }
 
 std::vector<double> CameraModelInitializeParams(const CameraModelId model_id,
