+++ src/colmap/sensor/models/radial.h
@@ -1,0 +1,203 @@
+// Copyright (c), ETH Zurich and UNC Chapel Hill.
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+//     * Neither the name of ETH Zurich and UNC Chapel Hill nor the names of
+//       its contributors may be used to endorse or promote products derived
+//       from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+#pragma once
+
+#include "colmap/sensor/models/base.h"
+
+namespace colmap {
+
+// Simple camera model with one focal length and one radial distortion
+// parameter.
+//
+// This model is similar to the camera model that VisualSfM uses with the
+// difference that the distortion here is applied to the projections and
+// not to the measurements.
+//
+// Parameter list is expected in the following order:
+//
+//    f, cx, cy, k
+//
+struct SimpleRadialCameraModel
+    : public BasePerspectivePinholeCameraModel<SimpleRadialCameraModel> {
+  PERSPECTIVE_CAMERA_MODEL_DEFINITIONS(CameraModelId::kSimpleRadial,
+                                       "SIMPLE_RADIAL",
+                                       "f, cx, cy, k",
+                                       1,
+                                       2,
+                                       1,
+                                       true)
+
+  static inline std::vector<double> InitializeParams(const double focal_length,
+                                                     const size_t width,
+                                                     const size_t height) {
+    return {focal_length, width / 2.0, height / 2.0, 0};
+  }
+
+  template <typename T>
+  static inline bool ImgFromCam(const T* params,
+                                const T& u,
+                                const T& v,
+                                const T& w,
+                                T* x,
+                                T* y,
+                                const bool check_cheirality = true) {
+    if (!HasProjectableDepth(w, check_cheirality)) {
+      return false;
+    }
+
+    const T f = params[0];
+    const T c1 = params[1];
+    const T c2 = params[2];
+
+    const T uu = u / w;
+    const T vv = v / w;
+
+    // Distortion
+    T du, dv;
+    Distortion(&params[3], uu, vv, &du, &dv);
+    *x = uu + du;
+    *y = vv + dv;
+
+    // Transform to image coordinates
+    *x = f * *x + c1;
+    *y = f * *y + c2;
+
+    return true;
+  }
+
+  static inline bool CamFromImg(
+      const double* params, double x, double y, double* u, double* v) {
+    const double f = params[0];
+    const double c1 = params[1];
+    const double c2 = params[2];
+
+    // Lift points to normalized plane
+    *u = (x - c1) / f;
+    *v = (y - c2) / f;
+
+    return IterativeUndistortion(&params[3], u, v);
+  }
+
+  template <typename T>
+  static inline void Distortion(
+      const T* extra_params, const T& u, const T& v, T* du, T* dv) {
+    const T k = extra_params[0];
+
+    const T u2 = u * u;
+    const T v2 = v * v;
+    const T r2 = u2 + v2;
+    const T radial = k * r2;
+    *du = u * radial;
+    *dv = v * radial;
+  }
+};
+
+// Simple camera model with one focal length and two radial distortion
+// parameters.
+//
+// This model is equivalent to the camera model that Bundler uses
+// (except for an inverse z-axis in the camera coordinate system).
+//
+// Parameter list is expected in the following order:
+//
+//    f, cx, cy, k1, k2
+//
+struct RadialCameraModel
+    : public BasePerspectivePinholeCameraModel<RadialCameraModel> {
+  PERSPECTIVE_CAMERA_MODEL_DEFINITIONS(
+      CameraModelId::kRadial, "RADIAL", "f, cx, cy, k1, k2", 1, 2, 2, true)
+
+  static inline std::vector<double> InitializeParams(const double focal_length,
+                                                     const size_t width,
+                                                     const size_t height) {
+    return {focal_length, width / 2.0, height / 2.0, 0, 0};
+  }
+
+  template <typename T>
+  static inline bool ImgFromCam(const T* params,
+                                const T& u,
+                                const T& v,
+                                const T& w,
+                                T* x,
+                                T* y,
+                                const bool check_cheirality = true) {
+    if (!HasProjectableDepth(w, check_cheirality)) {
+      return false;
+    }
+
+    const T f = params[0];
+    const T c1 = params[1];
+    const T c2 = params[2];
+
+    const T uu = u / w;
+    const T vv = v / w;
+
+    // Distortion
+    T du, dv;
+    Distortion(&params[3], uu, vv, &du, &dv);
+    *x = uu + du;
+    *y = vv + dv;
+
+    // Transform to image coordinates
+    *x = f * *x + c1;
+    *y = f * *y + c2;
+
+    return true;
+  }
+
+  static inline bool CamFromImg(
+      const double* params, double x, double y, double* u, double* v) {
+    const double f = params[0];
+    const double c1 = params[1];
+    const double c2 = params[2];
+
+    // Lift points to normalized plane
+    *u = (x - c1) / f;
+    *v = (y - c2) / f;
+
+    return IterativeUndistortion(&params[3], u, v);
+  }
+
+  template <typename T>
+  static inline void Distortion(
+      const T* extra_params, const T& u, const T& v, T* du, T* dv) {
+    const T k1 = extra_params[0];
+    const T k2 = extra_params[1];
+
+    const T u2 = u * u;
+    const T v2 = v * v;
+    const T r2 = u2 + v2;
+    const T radial = k1 * r2 + k2 * r2 * r2;
+    *du = u * radial;
+    *dv = v * radial;
+  }
+};
+
+}  // namespace colmap
